post_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
aiokafka
asyncio
python-jose[cryptography]
redis